Python/requirements.txt
pywin32>=306
Pillow>=10.0.0
requests>=2.31.0
numpy>=1.24.0
